+++ Program.cs
@@ -5,6 +5,7 @@ internal static class Program
     [System.STAThread]
     public static void Main()
     {
+        MainMenu.PrintInstructions();
         Game.Run();
     }
 }